Move smartctl warning to scenario
Adds a scenario to raise a warning when unhealthy disks
are reported by smartcl. Also make output none if nothing
to report.
